fix: risolvi duplicati post-merge HEL-501 (entity, repository, pom, yml, data.sql, Integer->Long)

+++ backend/src/main/java/com/aether/backend/dto/MissioneResponse.java
@@ -17,8 +17,8 @@ public class MissioneResponse {
     private String codice;
     private String obiettivo;
     private String descrizione;
-    private Integer idColonia;
-    private Integer idTipologia;
+    private Long idColonia;
+    private Long idTipologia;
     private OffsetDateTime dataInizioPrevista;
     private OffsetDateTime dataFinePrevista;
     private Priorita priorita;
