+++ backend/src/main/java/com/aether/backend/repository/TipologiaMissioneRepository.java
@@ -3,5 +3,5 @@ package com.aether.backend.repository;
 import com.aether.backend.entity.TipologiaMissione;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TipologiaMissioneRepository extends JpaRepository<TipologiaMissione, Integer> {
+public interface TipologiaMissioneRepository extends JpaRepository<TipologiaMissione, Long> {
 }
